pyproject.toml
[project]
name = "smartvision"
version = "0.1.0"
description = ""
authors = [
    {name = "tju",email = "tony.ju@me.com"}
]
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "oss2 (>=2.19.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "streamlit-chat-animated (>=0.0.3.3,<0.0.4.0)",
    "streamlit-chat (>=0.1.1,<0.2.0)",
    "streamlit-chatbot (>=0.0.2.2,<0.0.3.0)",
    "streamlit-option-menu (>=0.4.0,<0.5.0)",
    "streamlit-carousel (>=1.1.2,<2.0.0)",
    "streamlit (>=1.45.0,<2.0.0)",
    "ultralytics (>=8.3.128,<9.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "pynvim (>=0.5.2,<0.6.0)",
    "python-lsp-server[all] (>=1.12.2,<2.0.0)",
    "pylsp-mypy (>=0.7.0,<0.8.0)",
    "python-lsp-isort (>=0.2.1,<0.3.0)",
    "python-lsp-black (>=2.0.0,<3.0.0)",
    "llama-index-llms-deepseek (>=0.1.1,<0.2.0)",
    "llama-index-llms-ollama (>=0.5.4,<0.6.0)",
    "llama-index (>=0.12.34,<0.13.0)",
    "ollama (>=0.4.8,<0.5.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "jupyterlab (>=4.4.2,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
